Start the work of proxying the BGG game data
This commit scaffolds in the start of some code for an endpoint that can
look up the information for a game from the BoardGameGeek API and return
back the fields that we care about in a JSON format.

This is not all of the data that is provided, and also not all of the
data that we currently want, since some of it is in a nested structure
that is going to take a bit more work to parse.

This is also not the final production ready version of the code, which
will come once the data we want to gather has been stabilized.