Completed and cleaned up v1 of BGG API bridge
This finalizes the original API that we want to create for being able to
pull boardgame information from BGG in order to be able to ingest it
into our own database.

As a part of this we decided to punt initially on using the poll
information from the BGG data to try and scaffold out the list of
recommended ages and player counts, since it is a bit unclear how they
intend for us to interpret the data, and we've spent enough time on it
as it is.