Clean up and slightly refactor the BGG API
We realized late in stream yesterday that the BGG API doesn't return a
404 for games that don't exist, it returns a success message whose
internals tell you that it failed. I have no words (not true, I do have
words, but I don't want to type them here).

This cleans up the code to detect when this is happening so we can
return an appropriate error back.

In addition we noticed a small bug in which some things were testing
against null instead of undefined.