Implement the first parts of the game insert code
Implement the first bits of code that we can use to ingest game
information into the database using data from the existing BGG endpoint
we made.

There is still work to do here, and this needs to be cleaned up
somewhat, but it is a workable start.

This is a late check-in for day 4, for sleep reasons.