Clean up the initial game insertion code
This includes some simple Schema validation logic to ensure that the
game endpoint gets an object with the required fields in it, and that
any of the optional fields are filled out with default values.

Sadly this decouples the defaults from the DDL in the schema file, but
we can worry about that later.