Clarify the metadata update code
This refactors the code that we wrote for the metadata insert to make
it smarter.

We now predicate inserts on the slug and not on the BGGId since we want
to use this for all inserts potentialy, and not just games that come
from BoardGameGeek.

There is a new common routine which can determine which items need to
be inserted and insert only those. In addition, it returns back the
full database records for the items, inserts or not, so that their ID
values can be used. Thus this acts as an update and fetch all in one.