Update game inserts so that they add their meta
Using the power of the previous function that we added, this updates the
code for inserting a new game so that it ensures that all of its
metadata is up to date in the correct tables.

With this, the addition endpint can now fully flesh out a game based on
the data that falls out of the BGG Proxy Endpoint.

As a part of this change we also updated the schema, which was
referencing the wrong table and causing problems. The order of the
statements was also modified so that we can drop tables while there's
data without violating the foreign key constraints.