Refactor the BGG extraction code
In order to support our goal of an endpoint that can look up a game on
BGG and then insert the data directly into the database in one
operation, this refactors the BGG extraction code so that the core of
actually making the query is a reusable function, and the route handler
just invokes it.

We have also included a custom exception for this to make handling
of errors easier.