Swap the game metadata table sets for an ubertable
Previously we had 5 different tables to represent each of the different
types of metadata, and 5 other tables to track which of those are
associated with any one game.

This works (and is what versions 1 and 2 of this project did way back
in the day), but it seems messy and clunky.

This has been swapped to a system where there's a single table that
holds all of the metadata, and a new column tells you the type.

This seems to be just as fast, makes the data easier to look at, and it
reduces the complexity of the query needed to pull all of the data out.