Include an API to bulk insert games
After some refactoring on the code that does inserts based on a BGG Game
ID to make it reuable, a new endpoint was added that allows for bulk
inserting games by their BGG ID.

Using this, the entire current game collection from the tracking sheet
was successfully inserted, and the data was spat out into a SQL file to
save it, since it is a lot of data.

The full bulk insert took 15.5 seconds, most of which was the time to
fetch the data down from BGG.

As a part pf this change, the schema had to change because it turns out
that slugs overlap between different types of metadata.