Include a dump of the current game collection data
One of the last things that happened yesterday was we ran through the
list of all owned games in the collection and pulled down their data
into the local DB.

In order that we don't need to do that too frequently, we want to keep a
dump of known good data that we can insert back into the database.

We tried that with a whole database dump, but Wrangler seems a bit dumb
about how it works. So instead today we replace that one dump file with
many dump files, so it stops freaking out.

Also we no longer have explanatory comments in the schema because
Wrangler doesn't understand them.

Insert *sigh* emoji here.