Include an API to get metadata usage details
This implements a query by which it is possible to look up the
information on a metadata item based on its internal ID value or its
slug.

This will return information that gives you the name, slug and ID, and
can optionally also pull a list of games that are marked as using that
particular data.