Add DB metadata logging functionality
This wraps all of the calls that actually talk to the database so that
they report where they are making queries, what the query is doing, and
the number of rows read and written, as well as the result set size
(which may be different than reads), if any.

The idea is that we can use this to tune indexes on our queries by
seeing if we're scanning too much. Although it turns out that for
local miniflare, it's lame and doesn't provide those details.