Reorganize the current file layout
The set of planned changes is to extract out the code that is
responsible for the core of the database handling and put it into its
own file (or files).

In order to make that change cleaner we want to change the current
directory name, since it's overloaded right now but we want to leave
some of the content in its own separate folder.