Rename symbols to be less gross
As one of the final portions of this cleanup operation, this alters the
names of some symbols to be less dumb. The original names were the
result of an initial refactoring where we split single monolithin
functions for doing the full request into the DB wrapper and the requst
code.