Split the single requests file into two
To finalize the code reorganization, the single monolithic file whose
job it was to contain the requests has been split into two files so that
it is not so long and unwieldy.

Along the way we also tightened up the name of the metadata requests so
that they don't include the redundant "Game" portion, since all metadata
is currently by definition related specifically to games.