Create a single wrapper for all requests
Since all requests need to handle exceptions and use them to generate
error responses, instead of having every request be responsible for
that we now have a single wrapper function that can do that for us.

The single wrapper will take as an argument the function to invoke to
actually do the request, call it, and return what it returns. It also
handles all exceptions that might get thrown, and is the single point
at which any error handling for requests as a whole needs to occur.