Enhance the schema to start tracking ownerships
Include in the schema the tables needed to track the users of the
system, the households they live in, and the games that the household
currently owns.

This is data that may be interesting for the UI work that's upcoming if
we get to a point where we want to show distinct sets of information
instead of the data as a whole.