Out with old (ui), in with the new (ui)
After a brief ride on the struggle bus, this replaces the placeholder UI
that used Windstorm in favor of a new version that uses Zephyr instead.

The content attempts to be a fairly close match to what we had before,
though there may be some styling issues, and for the moment the button
doesn't work properly.