Implement some data table action for the game list
As a simple test of tying everything together, this implements a very
simple page which displays the list of games that are currently known
to the system.

This is very simplistic at the moment as we try to remind ourselves how
Svelte works in general, and as we try to decide what exactly all of the
desired pages are actually going to look like.