Clarify site navigation a bit more
To help scaffold out what it is that we're trying to build, this swaps
the crunky links at the top of the page for some sweet, sweet tabs
instead, and sets up so that all of the inner table links are able to
navigate to an appropriate page. Said page is currently just a
placeholder stub for the time being.