Fix a weird issue with hono routing
Not 100% sure, but it appears we found a bug in Hono, which trying to
update did not fix. This commit works around the issue however, so we
will stick with it for the time being.

The route for getting information on a specific metadata slug or ID type
was failing because Hono seemed to be grabbing the wrong part of the URL
for the parameter from the path.

A bit of investigation shows that this might be related to the argument
having one or more uppercase characters, even though metaType seems to
work OK.

Regardless, a quick swap to a different name allows it to work, so good
enough for now while we work on stuff.