Fix an improperly named identifier field
When gathering the data for games in a particular metadata subset, we
had a case of incorrect case for a field name in the game, gameID versus
gameId.

While this does not matter too much in the SQL end of things, it matters
a great deal to the JS code, since the column names become field names.