Refactor the SlugList component for flexibility
While implementing the next stage of the UI, we realized that the
structure of the data for games is represented differently
(but logically) at the API end because the optional list of games is in
a sub-field that has keys that more readily identify what the record is
for.

In order to be able to use the slug list for all of the cases that we
need, this needs to be more flexible.

To that end, the list now has a new prop names filter that allows you to
specify a function that will filter the results of the query such that
the slug table can contain it.

This is done in a backwards compatible way so as not to break any
existing implementation.