Provide some simple page titles for the lists
In order to make the site less inscrutable to humans, each of the lists
now displays a simple title along the top that represents what the
content of the list represents.

This could use some work, but at least for the time being it makes it
clearer what you're looking at.