Include a dump file that contains raw image URL's
In a minor data mining operation, this creates a simple dump file that
allows us to create a table that contains the underlying BGG images for
each of the games in the preloaded version of the database.

In order to gather this information we did a bulk query from BGG to
get the raw data for all games, so we have included that here.

Note that the raw dump doesn't contain the statistics information that
we originally used to gather the data.