Fix one last place where the game detail was wrong
In one of yesterday's commits, we fixed up some stuff regarding how the
game Too Many Cooks that we initially seeded was not the correct one,
since there is a newer game using that title and the older one is now
known by a different name.

In that commit we forgot to fix this file, which has the list of IDs
that constitute the games in the DB for seeding purposes.

So we're fixing that now to forestall any future sadness.