Upgrade the version of wrangler everywhere
Even though we use a globally installed version of Wrangler, this
updates to the most recent version so that at least anyone checking out
the code will have a hope of having the appropriate tool available.

The prompt for this change was hoping that the bug they introduced a
couple of weeks ago that leaves zombie workers blocking the worker port
might have been fixed, but alas, no.

You ever get that feeling you're the only one on an entire class of OS
to have an issue, based on it being so glaringly obvious?