Implement a solution to ingest BGG Game Images
Using information from the previous insertions of temporary images, this
endpoint can look up one or all of those records and insert into the
CF Images storage a version of that image for later retreival.

Images have names internally that namespace them based on the project
and asset type.

The URL of such images looks something like the following, noting that
when using our custom domain, the /images/ at the start comes from the
rule we use to proxy the URL to our domain.

https://imagedelivery.net/Dx4jrYPNAcY0IcNpYD9HJA/luddatumbazo/game/game1.jpg/public
https://luddatumbazo.odatnurd.net/images/luddatumbazo/game/game75.jpg/public