Include an imagePath in the game records
Although we want to have files that attach and images that attach to a
variety of data types, we want to ensure that all games have a native
image attached to them for their own display purposes in lists.

This augments the schema to include such a file; all data in the
database also had to have an appropriate URL injected, which we will be
working on doing automagically going forward.