Include a facility to map internal image paths
This commit implements a system by which one of our internal images can
be converted into a URL based on configured options and the purpose for
which the image is going to be put.

This uses an internal URI scheme of cfi:// to represent Cloudflare
images, since we plan to eventually have multiple back ends.

The routine can convert such a scheme into a proper URL based on the
configured environment variables.