Update the schema to support session reporting
Scaffold in some initial table structures that we can use to track game
sessions via session reporting.

This includes a simple "Guest" system to allow for people that aren't a
part of the system to participate in games, and also allows for us to
transition users into a guest, keeping their data while dropping their
ability to log in.