Remove the original tracking for game expansions
The original version of the code for tracking what games are expansions
to other games was broken in a variety of ways, not the least of which
being that it is a many to many situation and not a one to many like we
originally theorized.

Since the data would have been wrong if in fact anything had ever tried
to use it, we have removed the column that represents it in the
database.