Include a schema for tracking game expansions
What we have learned is that:
  - A BGG game entry includes a list of all expansions
  - A BGG expansion entry includes a list of all base games it expands
  - It is possible somehow for an expansion to itself be expanded

To that end we need a structure that will allow us to store a list of
all current and potential future expansions and base game entries so
that we can join them together when both parts are added without having
to have all possible games in the database.

This schema currently contain an entry that is the name provided in the
BGG query for the expansion/base record, so that we have something to
display if the matching game is not yet in the database.