Include a stub request for updating expansion data
Set up a test request, and the stub function it calls, so that we can
start working on the insert/patch of expansion data.

A comment that describes what should happen has been made; we should be
able to generalize the functionality such that it is not too
complicated since the same things happen for expansions and base games,
with only the side of the affected relation changing.