Start fleshing out expansion linking
Although not complete based on various issues that were encountered with
my lack of ability to remember what happens when you forget to await an
async call, this starts in on the task of defining how the endpoint that
will link games and their expansions together will work.

This needs to be further fleshed out and defined. 

This commit was also made late since I went to bed without actually 
pressing the commit button because I'm smart like that.