Update docs and enforce new input data constraints
This adjusts the documentation comment on the code that we started
working on for expansion handling to make it clearer how it expects to
work.

In addition, the input data is now properly validated against our
expectations so that all code that deals with the input knows for sure
that it is working with known good data.