Extend the expansion API to allow BGG inserts
This includes a small endpoint change that allows for inserting the
expansion information for a game based on it's BGG ID.

The endpoint will look up our game entry for that BGG ID and, if found,
look up the expansion information on BGG and then insert it using the
prior endpoint.

This saves effort and the potential mistakes of trying to construct all
of the entries that we need for the initial data population.