Include in the dump a list of our expanded games
As a part of the initial data dump structure, this stores a list of all
of the BGG Game ID's for games that are in the database for which there
is an expansion or for which that game IS an expansion.

The full state of the game expansions as it exists with our data is also
included for dump backup purposes.