Allow new game insertions to add/update expansions
This ties in the code for updating expansions so that it gets invoked
when games are inserted into the database as well.

This keeps things looking proper going forward.

This also resolves a bug that was introduced in the insertion code when
we removed the expandsGameId value from the game table; the insert was
not updated as regards the bound parameters and was failing.