Add a cleanup API for metadata and purge crud
Due to all of the times that we have accidentally inserted games due to
incorrect BGG mappings, this commit introduces an endpoint to find and
also (optionally) purge all metadata records of a specific type which
are not currently in use by anything.

This also cleans up the database somewhat by removing all of the entries
for deleted things from the inline dump and adding image records for
some games that did not have images because we had accidentally broken
the game insert during the removal of the expansion field.