Add images to game lists and details pages
This is a small but hard fought change to allow images for items in a
SlugList (where they exist) as well as showing the main box art image
on the game details page.

This is not exactly how we'd like it at the moment, but it's good enough
for now.