Include the list of expansions and base games
For better or worse, the game details pages now include links to all of
the expansions and base games (as appropriate) for any particular game.

Links are generated to internal pages if the game exists in our database
and to BGG (with an annotation) if not.

Currently there is something borked, possibly with svelte-spa-router,
which does not want us to link to other game pages (or rather, the URL
updates but the page does not load for some reason).

We will need to investigate that.