Fix an issue with HTML encoding in expansion names
When adding in the code to pull in BGG expansion information we forgot
that because they're lazy/using XML, their names have encoded HTML
entities in them that need to be decoded.

As such, this fixes that problem at the expense of yet another database
cleanup as we determine which expansions have dumb names now.