Fix our issue with navigating between games
We ran into an issue yesterday in which a game trying to link to some
other game did not properly reload the page.

This is because when the route changes the same top level page component
is still required, and since it is already mounted nothing else needs
to happen as far as Svelte is concerned.

The solution to this problem is to make the display portion a component
and give it a prop that tells is the slug, and then make the main page
use #key to force re-mount the component when the slug changes.