Start implementing the session guest API
To start our work on the session reporting API, this scaffolds out the
base of requests for dealing with guest users, since they may end up in
guest reports and we need a way to manipulate those records.

Currently only the guest list is complete and the others are broken
copy pastes stubs. More on them tomorrow.