Update the schema for Users and GuestUsers
For an upcoming change we need a single unique field (like a slug but
not) to uniquely identify users and guest users by a value that is not
the internal ID.

To do that we're using a generated column that gets the full name based
on the inner first and last names concatenated together.

At the same time we also realized that the order of the tables when
applying a dump was wrong because game owners references the household
and not the other way around.