Implement the request to get session details
This implements the set of queries that are needed to pull down all of
the information about the session report with a given ID and format it
into a JSON object.

This will likely be tweaked more, such as to make it batch all of the
queries for speed reasons, and to convert the incoming number values
into actual booleans.