Properly report session boolean fields as bools
D1 doesn't like when you give it booleans, but we want fields to
actually have boolean values.

We are thus going to follow on the theory that any field with name that
starts with "is" is a boolean field.

New routines can convert such fields to integers and back so that the
JSON conveys booleans but the values get stored as integers in the DB.