Implement an initial request to get a session list
This currently returns bare bones information about each session, which
does not include the expansions, the players, or any information on
them such as who won (not even a count of such games).

In addition we currently don't support any sort of criteria on the
search, which is also going to be interesting in the future.