Update the Session Report Schema data
This updates the session report table to allow for the session end time
to be a NULL value, which is an indication that the session is
currently still ongoing and has not finished yet.

This allows for using the app to log the start of a session, then play,
then come back and enter in the details without having to keep the page
open, which could cause issues.

Code will be made smart enough to know that a NULL session end means
that the session didn't close out.