Include the ability to get game synopsis details
Include a common DB helper that fetches for any given gameId the minimum
synopsis information, such as that used for a list of games or sublist
of game details.

This is useful for gathering such data in an ad-hoc fashion for single
games, such as when validating input and requiring the game data for
the output.