Start the work on a function for adding sessions
This is still a work in progress, but this starts the work that needs to
be done to allow us to add in a session report. We have the stubs of
some code, and the steps required for others already outlined, which
will need a bit of code effort to finish.