Update the game synopsis helper call signature
This updates the helper function that we use to look up the synopsis of
games so that instead of taking a single gameId it can also optionally
take an array of gameId's instead.

This allows the caller to look up a single or multiple games.

The return values remain the same for a single gameId, rather than doing
an adjustment that makes it always return an array, so as to keep all of
our backwards compatibility.