Finish the validation of session report data
When performing an insertion of a new session report entry, we need to
validate all of the incoming data and also perform some lookups in
order to know the data we need while inserting.

This change ensures that all data is valid and that we have all of the
fields that we need. All that remains is to do the insert, which we
will do next.