Add expansion name to session expansion details
This is a small change, but since for the game portion of a session
report we store the ID of the name of the game played so that the
report tracks the name we own the game by and not the primary name, the
same should be done for expansions.

It did not used to be so, but now that be like it is.

Also, there was a small typo in the mapping that populated the main game
ID, which was using the wrong field name for the source, which is now
fixed.