Modify the game lookup to take an image type
The common routine for looking up either games or expansions expands out
images to full paths; this has been modified so that it takes as an
argument the type of image, because sometimes you want a thumbnail and
sometimes you want a larger image, etc.

The call points for this currently are in the session data, where we
needed to distinguish these.