Alter validation of session report reporter
The intention was that the session reporter be one of the users that is
playing, since that field is tied to entries in the session report
player table.

This was not being done, and in fact it was possible to specify a
recording that would be validated but never stored; fixed now.