Finish up the session report insertion code
This ties all of our recent changes together by creating a function that
can do the inserts for a session report, and execute it as a part of the
insertion request.

The inserted data is reshaped so that it appears to be a regular lookup.

Additionally, as a test run, the insertion runs in a batch to make it
fully transactioned.