Implement the session update functionality
Update the session API's  so that it's possible to adjust the game
results and end time after the session was recorded.

This can be used either to "close out" an open session after the game is
complete, or to edit a previously added session when you realize you
boned it good.

This does not allow for fundamental changes to sessions; for that you
should delete and then re-create. The UI could have some way to capture
the data for an existing session and get ready to re-insert with
changes, for example.