Include a game filter in the session report list
This is a small update to the session list request API that allows you
to filter the list of reported sessions down such that it contains only
sessions for the games mentioned in the list.

The games specified can be in either the main game or the expansion
side of the session report.