Make the header buttons not touch the edges
Mobile browsers are apparently giant dicks in as much as the buttons in
the title bar look correct on the PC but are 1 pixel too tall on mobile
making the top of the button border get cut off and causing OCD
tendencies of epic proportions.

This includes some back magic math by @axel669 to adjust the margin
while also maintaining the aspect. Yay!