Include a simple session list page
This is some quick work to include a page that allows us to see a list
of all of the session reports that the system knows about.

We've included some columns to show whether or not the game is a
learning game, a link to the game page for those who care, and the
title. We also show the box art for the game, since we kinda already
had that in the code we copied and pasted from ourselves.

This is currently linking out to session report detail pages that don't
actually exist.