Include in game details results a session boolean
This adds a small subquery to the list of queries that are used to
gather game information so that we get an indication as to whether
there are any session reports for that game or not.

This doesn't return a list of them or even a count, just a boolean that
indicates that there is at least one.