Include a simple game lookup functionality
Since it seems like we may need this either for the UI, for tooling, or
both, this implements a lookup function that allows you to look up a
list of games based on their ID values or slugs (or a mix of both) and
get just the id and slug values.

This is used to validate input and/or perform matching without having
to fetch the whole of a game record to do it.