Fix our URL slug and ID routing
We were previously ham handedly trying to pluck the slug from the URL,
but actually the router provides automatic params to all components
that it routes that gives you the values of all of the named portions
of the route that landed at the component.

We now use that, which is less prone to fail when URL's have a harmless
trailing slash (for example).