Update the README to mention unused libraries
The original day-0 commit included a README with a list of potential
libraries that could be used throughout the project. We did not get as
far as using R2 for this project, so that library was redacted away
along with the ID library that we were going to use to make incoming
names unique.