Clean up the back navigation a little bit
The original navigation back buttons were navigating directly to
specific places rather than going back because originally they were
intended to take you back to specific places and not just where you
came from.

After doing more iterations on the UI, we have decided that this should
not work like that, so now the back is more explicit and uses a
specific new component so that all behave the same.