Factor the common BGGLink code out to a component
There are a few places where we were generating links to BGG Pages, and
they are all multi line and include tests to see if there is a BGG ID
at all.

This change wraps those into a simple component so that the page is
shorter and also more expressive regarding what it is actually doing.