Factor the image display code into a component
Similar to the BGGLink component we just created, this wraps all of the
images in a single component that does the checks to make sure that the
image path is provided and acts accordingly.

This will put all images related to games going through the same
component which will make future updates work better.