Include a README update for the new repo
This commit marks the point at which we split from devember-2023 and
went directly into the luddatumbazo repository; the README is updated
to suit.

Since I ask myself this question every year, no I did not fork the repo
because you can't fork a repo and keep it in the same user. So I just
altered the upstream URL that I push to and removed all of the day tags
and called it good.