Include some dump updates for teaching URLs
This syncs into the base data teaching URL's for all of the games that
have videos by Watch it Played on YouTube, which is our preferred source
for teaching.

In addition, there is also one or two others that are not related to
that channel to try and fill things out.