Sync in all of the current session reports
This brings in the initial version of all of the session reports dating
from Nov 2023 to the current time.

This required adding Tiny Towns to the database so that we could track
the play, so this is also added to the wishlist but not to the list of
owned games.