Include some indexes to speed up queries
In the cold light of day, many of the tables did not have indexes
(or proper ones) based on how miniflare can't tell you about reads and
writes the same as the production cloudflare can.

These small optimizations cut queries significantly where possible,
though some (such as pulling the entire game list) have significant
reads (to a degree) based purely on volume; they are currently as good
as they can get.

In addition to this, we also streamline a little bit some game lookups
to shave some reads to look try to look up the id values of games to
filter the session list if no values are provided for the filter, since
this ends up costing reads anyway.