Fix session report lookups
Now that there are more than just a couple session reports in the
database, it is clear that the original query had an unfortunate join
that stops it from being able to find session reports if there is a
filter in place. This also stops the query that tells game details when
there is a session report from working.

The fix (at least for now) is to search both tables for a hit and union
the lists.