Adjust the session report details page
As a test, this alters the details page for a session so that the date
and time of the session is displayed in a more readable way, the order
of the chips is different (since the learning game is optional), and
the layout of the image and description is modified such that they are
side by side and the text actually wraps.

This last change may not be kosher for phone screens, so we may need to
defer to an over-under like we do for games, but this is just a test.
Certainly we need to change how it lays out so that it wraps the text.