Reverse sort sessions by default
Now that there are not just a couple of session reports in the system,
modify things so that session reports can optionally be retrieved in
reverse sort order.

We also modify the client to request all session lists in reverse order
so that they show the most recent at the top.

The table has been modified so that it can sort the sessions in the list
based on the date. Currently this is just a duplicate of the natural
text sort order because the dates logically sort that way as ISO
strings since they all share the same time zone (i.e. none). 

We may want to convert the objects to dates, in which case we may need a
custom sort, so for the time being we will leave this as is.