Include session duration in details
As a test of what Luxon can do, incorporate a quick section in the page
that includes in the session timing information the duration over which
the game took.

This can be later extended to be a part of the session list as well, so
that we can see the duration of games in the list.