Embark upon an epic journey of data validation
This is the first test commit of a new scheme whereby all of our API
methods get full data validation. Currently that is in the form of zod,
although we have plans to swap that to something else like joker.

As a result, the changes here are designed to isolate the majority of
the code from the underlying mechanism, allowing us to do a simple
piecemeal replacement later as needed.

A part of this cleanup is for making the code a little more browseable,
readable and maintainable by splitting things up into smaller files and
having each section of the API have a distinct api route set, so that
the mainline code is smaller and simpler.

This batch of changes affects the guest user APIs specifically, as the
smallest of the API's currently.