Clean up the temporary image endpoint
This gives the cleanup treatment to the temporary image API we have that
allows bulk backfilling of images. The request now properly documents
the table that it expects to see.

This augments the asNumber() validation function to allow it to take
arguments, so that we can convert to optional numbers without it getting
angry at us.