Include a session report PlayType attribute
This change extends the SessionReport table and associated code so that
sessions can track what type of play the game was; physical game, BGA
game, steam, gog or android.

On the details page, the type of the play is displayed, and in session
tables an icon is displayed to give a hint as to the type(currently not
sortable or filterable).

This makes it possible for us to track not only physical game plays, but
also games from other sources (so long as we own them or they are in the
database).