Update Users and Guests to have a display name
This includes a new property on both guests and users alike that allows
for specifying a display name that is distinct and unique from the
actual name fields that identify the records.

For Users, this must be set explicitly when they're added currently
(which is a core DB action because there's no API for it yet).

For Guests, when manually updating guests it's possible to provide a
displayName for them, to be inserted. When guests are implicitly added
as a result of them being used in a session report with an explicit add,
the displayName is instead manually inferred, and can be changed later.

Such a change is also currently a core DB action, because although there
is an endpoint that allows for updating guests, it only adds guests
that do not exist, it does not allow editing the details of an existing
guest, since previously that was not allowed.