Fix an issue with zod error handling
In cases where the entire message body is of the wrong type (e.g. it
expects an array but gets an object, or vice versa), an server error
would be thrown due to an incorrect assumption of the shape of the zod
error object.

This has now been fixed so that we can catch that and report on such
problems properly instead of just blowing up.