Rearrange the code for metadata and expansions
This alters the structure of the code that is used to handle the
metadata and expansion parts of the game data to pull them into a
singular structure like that of previous commits.

Unlike those commits, these are not yet guarded by schemas properly as
we realize that the layout is starting to get a little hairy re: imports
and we pause to see if we can't make that situation less gross.