Upgrade hono to get fixes for their path errors
During devember we noted that if a route contains multiple parameters
and more than one of them has upper case characters, the returned value
for the parameter is the value of the wrong part of the path segment,
so we renamed one of the fields to get things working.

With the schema validation work we're doing, we do not want that to be
the case. This updates to the latest version of Hono, which it turns out
actually resolves the issue for us.