Update the metadata requests for zod validation
To finalize our initial work on validation (pending some additional
cleanup of things), this fleshes out the metadata requests to have full
schema support.

As a part of this, the purge operation for metadata was split such that
instead of a query argument, the method itself tells the request if it
should actually perform the purge or not.