Clean up common DB code
There were some routines remaining from the original iteration of the
code where we were doing minimal validation of input data that were
kicking around and needed to be removed, which is now done.

This also touched the session update code, which is the last part of the
older code that was still using that. This pointed out that we
accidentally left the update short circuited, but also that the update
schema should not have default values because they are meant to be
inherited from an existing session.